+++ api/src/main/java/com/edrdog/api/auth/dto/LoginRequest.java
@@ -1,4 +1,4 @@
-package com.edrdog.api.auth.web;
+package com.edrdog.api.auth.dto;
 
 /** 로그인 요청. */
 public record LoginRequest(String email, String password) {
