+++ api/src/main/java/com/edrdog/api/auth/dto/MeResponse.java
@@ -1,4 +1,4 @@
-package com.edrdog.api.auth.web;
+package com.edrdog.api.auth.dto;
 
 /** 내 정보(/me) 응답. */
 public record MeResponse(Long userId, Long tenantId, String email, String role) {
