+++ api/src/main/java/com/edrdog/api/auth/service/AuthResult.java
@@ -1,4 +1,4 @@
-package com.edrdog.api.auth;
+package com.edrdog.api.auth.service;
 
 /**
  * 회원가입/로그인 성공 결과. 발급된 세션 토큰과 현재 유저 정보를 담는다.
