+++ api/src/main/java/com/edrdog/api/auth/service/Principal.java
@@ -1,4 +1,4 @@
-package com.edrdog.api.auth;
+package com.edrdog.api.auth.service;
 
 /**
  * 토큰으로 확인된 현재 유저(토큰 제외). resolve 결과에 쓰인다.
